+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 * inject a 'SOXSPIPE PRE' keyword with timestamp value into prepared frames
 * check frames for 'SOXSPIPE PRE' keyword before preparing - raises exception if found
+* moved basic input frame verifications to the `_base_recipe` - so not to repeat code
+* added mixing of readout speeds to input frame verification checks
 
 ## v0.2.0 - February 27, 2020